Removing initial Friendlinks
Since the accounts aren't set up, lets put announcing them on hold. We can add them one at a time and use that as a way to have a news announcement for each one.  We can trickle one out every couple weeks or so. So there is a steady stream of news of some variety.